fix: keep confirmed-dead latch and split runtime tests
failNotConfirmed no longer clears confirmedDead so overlapping Kill/Drain
timeouts cannot undo a concurrent confirmation (reusable PGID guard).
Move retry/pipe/platform runtime cases into containment_runtime_retry_test.go
so no single subsystem test file exceeds the 300-line guideline.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
